+++ tesys/src/astrometry/src/lib.rs
@@ -1,12 +1,12 @@
 extern crate chrono;
 
 // Constants
-const DEG_PER_RAD: f32 = { 180.0 / std::f32::consts::PI };
-const HOUR_PER_DEG: f32 = { 1.0 / 15.0 };
-const HOUR_PER_RAD: f32 = { HOUR_PER_DEG * DEG_PER_RAD };
+const DEG_PER_RAD: f64 = { 180.0 / std::f64::consts::PI };
+const HOUR_PER_DEG: f64 = { 1.0 / 15.0 };
+const HOUR_PER_RAD: f64 = { HOUR_PER_DEG * DEG_PER_RAD };
 
-const DEFAULT_WRAP_MAX_ANGLE: f32 = 2.0 * std::f32::consts::PI;
-const DEFAULT_WRAP_MIN_ANGLE: f32 = 0.0;
+const DEFAULT_WRAP_MAX_ANGLE: f64 = 2.0 * std::f64::consts::PI;
+const DEFAULT_WRAP_MIN_ANGLE: f64 = 0.0;
 
 pub mod datetime;
 
@@ -28,7 +28,7 @@ pub use location::Location;
 mod skycoordinate;
 pub use skycoordinate::SkyCoordinate;
 
-pub fn wrap_angle(val: f32) -> f32 {
+pub fn wrap_angle(val: f64) -> f64 {
     // This is used to wrap the angle between the limits specified within the code.
     let mut rads = if val < DEFAULT_WRAP_MIN_ANGLE {
         DEFAULT_WRAP_MIN_ANGLE - (val % DEFAULT_WRAP_MIN_ANGLE)
